delete query.js
