--- gcp-serverless/query.js
@@ -1,27 +1,0 @@
-'use strict'
-
-module.exports = `
-query DashboardQuery($searchstring: String!) {
-    search(query: $searchstring, type: REPOSITORY, first: 100) {
-      edges {
-        node {
-         ... on Repository {
-           nameWithOwner,
-           ... RepoStatus
-         }
-       }
-      }
-    }
-}
-fragment RepoStatus on Repository  {
-  defaultBranchRef {
-    target {
-      ... on Commit {
-        status {
-            state
-          }
-      }
-    }
-  }
-}
-  `
